Move tests for new module `pyam.str` to own file
